requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
chardet==5.2.0
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
loguru==0.7.3
Markdown==3.7
markdown2==2.5.3
openai==1.61.1
pdfkit==1.0.0
pillow==11.1.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pyodbc==5.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
reportlab==4.3.0
sniffio==1.3.1
tqdm==4.67.1
wkhtmltopdf==0.2